+++ Steepfile
@@ -40,7 +40,6 @@ target :lib do
   check 'lib/procon_bypass_man/support/on_memory_cache'
   check 'lib/procon_bypass_man/support/uptime.rb'
   check 'lib/procon_bypass_man/support/safe_timeout'
-  check 'lib/procon_bypass_man/support/update_remote_pbm_action_status_http_client'
   check 'lib/procon_bypass_man/support/compress_array'
 
   signature 'sig'
