+++ lib/procon_bypass_man.rb
@@ -31,7 +31,7 @@ require_relative "procon_bypass_man/support/on_memory_cache"
 require_relative "procon_bypass_man/support/http_client"
 require_relative "procon_bypass_man/support/report_http_client"
 require_relative "procon_bypass_man/support/remote_macro_http_client"
-require_relative "procon_bypass_man/support/update_remote_pbm_action_status_http_client"
+require_relative "procon_bypass_man/support/update_remote_pbm_job_status_http_client"
 require_relative "procon_bypass_man/support/send_device_stats_http_client"
 require_relative "procon_bypass_man/support/procon_performance_http_client"
 require_relative "procon_bypass_man/support/analog_stick_hypotenuse_tilting_power_scaler"
@@ -55,7 +55,6 @@ require_relative "procon_bypass_man/device_model"
 require_relative "procon_bypass_man/procon/button"
 require_relative "procon_bypass_man/procon/analog_stick_cap"
 require_relative "procon_bypass_man/procon/analog_stick_manipulator"
-require_relative "procon_bypass_man/remote_pbm_action/value_objects/remote_pbm_action_object"
 require_relative "procon_bypass_man/scheduler"
 require_relative "procon_bypass_man/plugins"
 require_relative "procon_bypass_man/worker"
@@ -63,7 +62,6 @@ require_relative "procon_bypass_man/websocket/client"
 require_relative "procon_bypass_man/websocket/watchdog"
 require_relative "procon_bypass_man/websocket/forever"
 
-require_relative "procon_bypass_man/remote_pbm_action"
 require_relative "procon_bypass_man/remote_action"
 
 STDOUT.sync = true
