+++ lib/procon_bypass_man/remote_action/remote_pbm_job/reboot_os_action.rb
@@ -1,0 +1,24 @@
+module ProconBypassMan
+  module RemoteAction
+    module RemotePbmJob
+      class RebootOsAction < BaseAction
+
+        def action_content(_args)
+          ProconBypassMan.logger.info "execute RebootOsAction!"
+          `reboot`
+        end
+
+        private
+
+        def before_action_callback
+          ProconBypassMan::ReportStartRebootJob.perform
+          be_processed
+        end
+
+        def after_action_callback
+          # no-op
+        end
+      end
+    end
+  end
+end
