+++ lib/procon_bypass_man/remote_action/remote_pbm_job/report_procon_status.rb
@@ -1,0 +1,25 @@
+module ProconBypassMan
+  module RemoteAction
+    module RemotePbmJob
+      class ReportProconStatusAction < BaseAction
+
+        def action_content(_args)
+          ProconBypassMan.logger.info "execute ReportProconStatusAction!"
+          ProconBypassMan::SendInfoLogCommand.execute(
+            message: ProconBypassMan::ProconDisplay::Status.instance.current.to_s
+          )
+        end
+
+        private
+
+        def before_action_callback
+          be_in_progress
+        end
+
+        def after_action_callback
+          be_processed
+        end
+      end
+    end
+  end
+end
