--- lib/procon_bypass_man/remote_pbm_action/commands/run_remote_pbm_action_dispatch_command.rb
@@ -1,23 +1,0 @@
-class ProconBypassMan::RunRemotePbmActionDispatchCommand
-  # @param [String] action
-  # @param [String] uuid
-  # @return [void]
-  def self.execute(action: , uuid: , job_args: )
-    case action
-    when ProconBypassMan::RemotePbmAction::ACTION_CHANGE_PBM_VERSION
-      ProconBypassMan::RemotePbmAction::ChangePbmVersionAction.new(pbm_job_uuid: uuid).run!(job_args: job_args)
-    when ProconBypassMan::RemotePbmAction::ACTION_STOP_PBM
-      ProconBypassMan::RemotePbmAction::StopPbmAction.new(pbm_job_uuid: uuid).run!(job_args: {})
-    when ProconBypassMan::RemotePbmAction::ACTION_REBOOT_OS
-      ProconBypassMan::RemotePbmAction::RebootOsAction.new(pbm_job_uuid: uuid).run!(job_args: {})
-    when ProconBypassMan::RemotePbmAction::ACTION_RESTORE_SETTING
-      ProconBypassMan::RemotePbmAction::RestorePbmSettingAction.new(pbm_job_uuid: uuid).run!(job_args: job_args)
-    when ProconBypassMan::RemotePbmAction::ACTION_REPORT_PORCON_STATUS
-      ProconBypassMan::RemotePbmAction::ReportProconStatusAction.new(pbm_job_uuid: uuid).run!(job_args: {})
-    else
-      raise "#{action}は対応していないアクションです"
-    end
-  rescue ProconBypassMan::RemotePbmAction::ActionUnexpectedError => e
-    ProconBypassMan::SendErrorCommand.execute(error: e)
-  end
-end
