--- lib/procon_bypass_man/remote_pbm_action/restore_pbm_setting.rb
@@ -1,41 +1,0 @@
-module ProconBypassMan
-  module RemotePbmAction
-    class RestorePbmSettingAction < BaseAction
-
-      def action_content(args: )
-        require "pbmenv"
-        ProconBypassMan.logger.info "execute RestorePbmSettingAction!"
-        setting = args.dig("setting") or raise(ProconBypassMan::RemotePbmAction::NeedPbmVersionError, "settingが必要です, #{args.inspect}")
-
-        # 復元に失敗したら戻せるように退避する
-        FileUtils.copy(
-          ProconBypassMan::ButtonsSettingConfiguration.instance.setting_path,
-          ProconBypassMan.fallback_setting_path,
-        )
-
-        ProconBypassMan::YamlWriter.write(
-          path: ProconBypassMan::ButtonsSettingConfiguration.instance.setting_path,
-          content: setting,
-        )
-
-        hot_reload!
-      end
-
-      private
-
-      def before_action_callback
-        be_in_progress
-      end
-
-      def after_action_callback
-        be_processed
-      end
-
-      # @return [void]
-      def hot_reload!
-        Process.kill(:USR2, ProconBypassMan.pid)
-      end
-    end
-  end
-end
-
