+++ lib/procon_bypass_man/support/update_remote_pbm_job_status_http_client.rb
@@ -1,5 +1,5 @@
 module ProconBypassMan
-  class UpdateRemotePbmActionStatusHttpClient < HttpClient
+  class UpdateRemotePbmJobStatusHttpClient < HttpClient
     def put(to_status: )
       super(request_body: {
         body: { status: to_status },
