+++ spec/lib/procon_bypass_man/remote_pbm_action/commands/update_remote_pbm_job_status_command_spec.rb
@@ -1,7 +1,6 @@
 require "spec_helper"
 
-
-describe ProconBypassMan::UpdateRemotePbmActionStatusCommand do
+describe ProconBypassMan::UpdateRemotePbmJobStatusCommand do
   describe '.execute!' do
     let(:post_body) { "hoge" }
