+++ spec/lib/procon_bypass_man/remote_pbm_action/stop_pbm_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-describe ProconBypassMan::RemotePbmAction::StopPbmAction do
+describe ProconBypassMan::RemoteAction::RemotePbmJob::StopPbmJob do
   describe '#run!' do
     let(:action) { described_class.new(pbm_job_uuid: "a") }
     subject { action.run!(job_args: {}) }
